refactor: clarify business logic boundaries in comments modules
Extract Telescope pickers into comments/pickers.lua, move pure state
accessors to data.lua, and route all GitHub API calls through sync.lua.
This reduces comments.lua from 675 to ~280 lines with clear separation:
data (pure), sync (API), pickers (Telescope), facade (commands).
